Move calling templates to the source file, use original name of module files, change address LUT to two dimensions
